https

+++ ckj_13_mvc/src/main/java/com/itstep/controller/NoteController.java
@@ -1,5 +1,6 @@
 package com.itstep.controller;
 
+import java.security.Principal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Comparator;
@@ -17,71 +18,109 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import com.itstep.model.Note;
+import com.itstep.model.User;
 import com.itstep.repository.NoteRepository;
+import com.itstep.repository.UserRepository;
 
 @Controller
 @RequestMapping("/notes")
 public class NoteController {
 	// private List<Note> notes = new ArrayList<>();
 	private NoteRepository noteRepository;
+	private UserRepository userRepositoty;
 
-	@Autowired
-	public NoteController(NoteRepository noteRepository) {
-		this.noteRepository = noteRepository;
-	}
+	
+	
 
 	@GetMapping("/create")
 	public String create() {
 		return "create_note";
 	}
+	
+	@Autowired
+	public NoteController(NoteRepository noteRepository, UserRepository userRepositoty) {
+		this.noteRepository = noteRepository;
+		this.userRepositoty = userRepositoty;
+	}
 
 	@PostMapping("/add")
-	public String add(@ModelAttribute(name = "note") Note note) {
+	public String add(@ModelAttribute(name = "note") Note note, Principal principal) {
+	    User user = userRepositoty.findByUsername(principal.getName());
+	    
 		note.setZd(LocalDateTime.now().toString());
+		user.addNote(note);
 		noteRepository.save(note);
+		userRepositoty.save(user);
 		return "redirect:/notes";
 	}
 
 	@GetMapping
-	public String all(Model model) {
-		model.addAttribute("notes", noteRepository.findAll());
+	public String all(Model model,Principal principal) {
+		 User user = userRepositoty.findByUsername(principal.getName());
+		List<Note> notes= user.getNotes();
+		 model.addAttribute("notes", user.getNotes());
+		
 		return "notes";
 	}
 
 	@GetMapping("/{id}")
-	public String info(@PathVariable(name = "id") int id, Model model) {
-
-		Note note = noteRepository.findById(id).get();
+	public String info(@PathVariable(name = "id") int id, Model model,Principal principal) {
+		 User user = userRepositoty.findByUsername(principal.getName());
+		List<Note> notes = user.getNotes();
+		 Note note = noteRepository.findById(id).get();
+		 if(notes.contains(note)) {
 		model.addAttribute("note", note);
-
 		return "info";
+		 }
+		 else {
+			 return "redirect:/notes";
+		 }
 	}
 
 	@GetMapping("/delete/{id}")
-	public String delete(@PathVariable(name = "id") int id) {
-		noteRepository.deleteById(id);
-		return "redirect:/notes";
+	public String delete(@PathVariable(name = "id") int id,Principal principal) {
+		 User user = userRepositoty.findByUsername(principal.getName());
+		 List<Note> notes = user.getNotes();
+		 Note note = noteRepository.findById(id).get();
+		 if(user.getNotes().contains(note)) {
+		 noteRepository.deleteById(id);
+		 return "redirect:/notes";
+		 }
+		 else {
+			 return "redirect:/notes";
+		 }
+		
 	}
 
 	@GetMapping("/edit/{id}")
-	public String edit(@PathVariable(name = "id") int id, Model model) {
-		// model.addAttribute("note", noteRepository.findById(id).get());
-		Note note = noteRepository.findById(id).get();
+	public String edit(@PathVariable(name = "id") int id, Model model,Principal principal) {
+		 User user = userRepositoty.findByUsername(principal.getName());
+		 List<Note> notes = user.getNotes();
+		 Note note = noteRepository.findById(id).get();
+		 if(user.getNotes().contains(note)) {
 		model.addAttribute("note", note);
 		return "edit";
+		 }
+		 else {
+			 return "redirect:/notes";
+		 }
 	}
 
 	@PostMapping("/edit/{id}")
-	public String editNote(@ModelAttribute Note note, @PathVariable(name = "id") int id) {
+	public String editNote(@ModelAttribute Note note, @PathVariable(name = "id") int id,Principal principal) {
 		// String s = LocalDateTime.now().toString();
+		 User user = userRepositoty.findByUsername(principal.getName());
 		note.setZd(LocalDateTime.now().toString());
+		note.setUser(user);
 		noteRepository.save(note);
 		return "redirect:/notes/";
 	}
 
 	@GetMapping("/search")
-	public String search(@RequestParam(name = "word") String word, Model model) {
-		List<Note> notes = noteRepository.search("%" + word + "%");
+	public String search(@RequestParam(name = "word") String word, Model model, Principal prl) {
+		 User user = userRepositoty.findByUsername(prl.getName());
+		 word="%" + word + "%";
+		List<Note> notes = noteRepository.search(word, user.getId());
 		model.addAttribute("notes", notes);
 		return "notes";
 	}
